Split out the UUID management
