+++ opencog/persist/sql/multi-driver/CMakeLists.txt
@@ -8,6 +8,7 @@ ADD_LIBRARY (persist-sql
 	SQLBulk
 	SQLValues
 	SQLTypeMap
+	SQLUUID
 	SQLPersistSCM
 )
